Refactor SearchMoviesResults
