+++ src/features/SearchResults/SearchMoviesResults/api.js
@@ -1,0 +1,18 @@
+import axios from 'axios';
+
+const API_KEY = 'eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIxZDQ2MTI2MDcwMmJkZGJiOTg4MmUyZTRhMDJlZDA0ZSIsIm5iZiI6MTcyMzY2MTY0My45NDk0MjksInN1YiI6IjY2YWI4MjEwNGZlNDIxMzEwY2QyY2FlNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.9nMPyeWUKdhaUTVPfltvi473upjSJlz1iKauLIeuXpQ';
+
+export const fetchMovies = async (query) => {
+    const options = {
+        method: 'GET',
+        url: 'https://api.themoviedb.org/3/search/movie',
+        params: { query, include_adult: 'false', language: 'en-US', page: '1' },
+        headers: {
+            accept: 'application/json',
+            Authorization: `Bearer ${API_KEY}`
+        }
+    };
+
+    const response = await axios.request(options);
+    return response.data.results;
+};
